backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=2.0.0
langdetect>=1.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.0.0
